refactor API into a separate crate with traits
